requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
httpx==0.27.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
